AzureResourceManagerService/requirements.txt
jmespath~=0.10.0
PyJWT~=1.7.1
cryptography~=3.1
requests~=2.24.0
adal~=1.2.4
python-dateutil~=2.8.1
six~=1.15.0
PyYAML~=5.3.1
knack~=0.7.2
msal~=1.0.0
humanfriendly~=8.2
colorama~=0.4.3
pkginfo~=1.5.0.1
argcomplete~=1.12.0
msrest~=0.6.19
msrestazure~=0.6.4
portalocker~=1.7.1
applicationinsights~=0.11.9
urllib3~=1.25.10
oauthlib~=3.1.0
isodate~=0.6.0
certifi~=2020.6.20
chardet~=3.0.4
cffi~=1.14.2
paramiko~=2.7.2
bcrypt~=3.2.0
Pygments~=2.7.0
idna~=2.10
pycparser~=2.20